requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.7
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
bleach==3.1.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-simple-email-confirmation==0.70
djangorestframework==3.11.0
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
msgpack-python==0.5.6
Pillow==7.0.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyfcm==1.4.7
PyHamcrest==2.0.0
pylint==2.4.4
pyOpenSSL==19.1.0
pytz==2019.3
redis==2.10.6
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.11.2
zope.interface==4.7.1
